NC archived_scripts
Add archived_scripts folder. I'd like to move scripts into here that we don't seem to be using anymore. After you add the EPSILOMETER file path to Matlab, remove this one and see if everything still works. Slowly, we can get to a point where all the scripts in EPSILOMETER are current.