updated epsiAuto_convert_raw_to_mat_and_plot to initialize epsi_class
We used to be limited to calling epsi_class only from the current directory, so there were many steps in this script that made something that looked like an epsi_class so that we could use epsi_class functions.

Now we can initialize an epsi_class in another directory, so we do that here.