epsi_class: Changed errors to warning and return
epsi_class: Changed errors to warning and return. That way, you can initialize an epsi_class to have access to the functions even if the data aren't in the right directory format. You can fill in the data later. (This is useful for mooring data where there is no attached SBE)